utils/trint/requirements.txt
docopt==0.6.2
logzero==1.5.0
webvtt-py==0.4.3
